Move `HadesLogsMixin` legacy test base out of `tests.__init__`
