+++ tests/frontend/legacy_base.py
@@ -152,3 +152,22 @@ class FrontendWithAdminTestBase(FrontendDataTestBase, FactoryDataTestBase):
         admin_group = AdminPropertyGroupFactory()
         MembershipFactory.create(user=self.admin, group=admin_group)
         self.config = ConfigFactory()
+
+
+class InvalidateHadesLogsMixin(testing.TestCase):
+    """Mixin Class forcing a disabled `hades_logs` extensions
+
+    This mixin class hooks into :meth:`create_app` and invalidates a
+    possibly configured `hades_logs` extension.  Useful if the default
+    is :py:cls:`DummyHadesLogs`.
+    """
+    def __init__(self, *a, **kw):
+        warnings.warn('Use pytest with the `session` fixture instead', DeprecationWarning)
+        super().__init__(*a, **kw)
+
+    def create_app(self):
+        app = super().create_app()
+        # invalidate already configured hades_logs
+        if 'hades_logs' in app.extensions:
+            app.extensions.pop('hades_logs')
+        return app
